+++ internal/global/instruments_test.go
@@ -146,7 +146,6 @@ type testCountingFloatInstrument struct {
 	count int
 
 	instrument.Float64Observable
-	instrument.Synchronous
 }
 
 func (i *testCountingFloatInstrument) observe() {
@@ -163,7 +162,6 @@ type testCountingIntInstrument struct {
 	count int
 
 	instrument.Int64Observable
-	instrument.Synchronous
 }
 
 func (i *testCountingIntInstrument) observe() {